requirements.txt
Flask==2.0.1
Flask-PyMongo==2.3.0
pytest==6.2.4